// src/app/shared/interfaces/poor-prognostic-factor.ts
export class PoorPrognosticFactor {
    age: number;
    imc: number; // poids(kg)/taille^2
    heartDisease: boolean;
    diabetes: boolean;
    cancer: boolean;
    breathingIllness: boolean;
    chronicRenalFailure: boolean;
    chronicLiverDisease: boolean;
    pregnancy: boolean;
    immuneSystemDisease: boolean;
    immunosuppressiveTherapy: boolean;
}